Replace exact matching code with ngram-based matching
The original exact matching code took in two texts to match and
performed some kind of alignment of the two texts prior to matching.
This alignment code was pretty opaque, not really documented, and very
messy.  In addition, if you try to run it, the swalign API that was used
appears to have changed -- some of the functions used no longer exist,
such as alignment.match(). That library is also designed to operate on
characters, not words, so it may not have been the right choice of
library anyway. There's some weird stuff later on when matching
substrings based on the alignment that tries to handle words that got
broken up (e.g., matching "frog dog w" from "frog dog woke" and "frog
dog water").

Thus, I've removed the original alignment code in favor of a simpler
form of ngram-based phrase matching, for ngrams of length N. While this
matching will produce duplicate matches for phrases longer than length
N, it still generates the general results we're looking for, i.e.,
higher match scores for texts that have more phrases, and more longer
phrases, in common.

Also:
- Add information about matching to the README.
- Update print lines in the main function to be more informative.
- Part of #4 and #5.