Replace similar phrase matching with ngram-based version
As mentioned in a68c2b76, the original code for matching phrases in two
texts performed some kind of alignment and was pretty messy. Here, I
remove the similar phrase matching and replace it with a simpler
ngram-based fuzzy matching, in line with how the exact matching was
replaced with an ngram-based algorithm. This will produce different
results than the original, but is more straightforward and
understandable, and still generates the general results we want: i.e.,
higher matching scores for texts that are more similar.

While here, I also add some new similarity metrics for the two texts,
such as the length, number of unique words, and overall fuzzy string
matching scores using several different ratios.

In sum:
- Remove old similar phrase matching and all its related functions.
- Remove unused imports.
- Add new similarity matching function.
- Add overall text similarity function.
- Add info on the overall text similarity metrics to the README.
- Add info on the similarity phrase matching to the README.
- While there, wrap some lines in the README.
- Fix #4, fix #5.